stream_processing/requirements.txt
clickhouse-driver
joblib
kafka-python
pandas
python-dotenv
xgboost
